Refining query mechanism and update map data

+++ backend/stuTravel/src/main/java/org/twentyEight/pojo/Venue.java
@@ -4,12 +4,12 @@ import lombok.Data;
 
 @Data
 public class Venue {
-    private Integer id;
+    private Long id;
     private String name;
     private String type;
     private Double latitude;
     private Double longitude;
-    private Integer placeId; // 外键
+    private Long placeId; // 外键
     private Integer popularity;
     private Long osmid;
 }
